+++ src/config/buildpack_config.rs
@@ -6,7 +6,9 @@ use indexmap::IndexSet;
 use toml_edit::{DocumentMut, TableLike};
 
 use crate::config::{ParseRequestedPackageError, RequestedPackage};
-use crate::{BuildpackResult, DebianPackagesBuildpackError};
+use crate::DebianPackagesBuildpackError;
+
+pub(crate) const NAMESPACED_CONFIG: &str = "com.heroku.buildpacks.deb-packages";
 
 #[derive(Debug, Default, Eq, PartialEq)]
 pub(crate) struct BuildpackConfig {
@@ -14,11 +16,14 @@ pub(crate) struct BuildpackConfig {
 }
 
 impl BuildpackConfig {
-    pub(crate) fn exists(config_file: impl AsRef<Path>) -> BuildpackResult<bool> {
-        Ok(config_file
-            .as_ref()
-            .try_exists()
-            .map_err(|e| ConfigError::CheckExists(config_file.as_ref().to_path_buf(), e))?)
+    pub(crate) fn is_present(config_file: impl AsRef<Path>) -> Result<bool, ConfigError> {
+        match BuildpackConfig::try_from(config_file.as_ref().to_path_buf()) {
+            Ok(_) => Ok(true),
+            Err(ConfigError::ParseConfig(_, ParseConfigError::MissingNamespacedConfig)) => {
+                Ok(false)
+            }
+            Err(e) => Err(e),
+        }
     }
 }
 
@@ -26,11 +31,8 @@ impl TryFrom<PathBuf> for BuildpackConfig {
     type Error = ConfigError;
 
     fn try_from(value: PathBuf) -> Result<Self, Self::Error> {
-        fs::read_to_string(&value)
-            .map_err(|e| ConfigError::ReadConfig(value.clone(), e))
-            .and_then(|contents| {
-                BuildpackConfig::from_str(&contents).map_err(|e| ConfigError::ParseConfig(value, e))
-            })
+        let contents = read_config_file(&value)?;
+        BuildpackConfig::from_str(&contents).map_err(|e| ConfigError::ParseConfig(value, e))
     }
 }
 
@@ -38,25 +40,9 @@ impl FromStr for BuildpackConfig {
     type Err = ParseConfigError;
 
     fn from_str(contents: &str) -> Result<Self, Self::Err> {
-        let doc = DocumentMut::from_str(contents).map_err(Self::Err::InvalidToml)?;
-
-        // the root config is the table named `[com.heroku.buildpacks.deb-packages]` in project.toml
-        let root_config_item = doc
-            .get("com")
-            .and_then(|item| item.as_table_like())
-            .and_then(|com| com.get("heroku"))
-            .and_then(|item| item.as_table_like())
-            .and_then(|heroku| heroku.get("buildpacks"))
-            .and_then(|item| item.as_table_like())
-            .and_then(|buildpacks| buildpacks.get("deb-packages"));
-
-        match root_config_item {
-            None => Ok(BuildpackConfig::default()),
-            Some(item) => item
-                .as_table_like()
-                .ok_or(Self::Err::WrongConfigType)
-                .map(BuildpackConfig::try_from)?,
-        }
+        let doc = parse_config_toml(contents)?;
+        let config = get_buildpack_namespaced_config(&doc)?;
+        BuildpackConfig::try_from(config)
     }
 }
 
@@ -70,7 +56,7 @@ impl TryFrom<&dyn TableLike> for BuildpackConfig {
             for install_value in install_values {
                 install.insert(
                     RequestedPackage::try_from(install_value)
-                        .map_err(Self::Error::ParseRequestedPackage)?,
+                        .map_err(|e| Self::Error::ParseRequestedPackage(Box::new(e)))?,
                 );
             }
         }
@@ -81,7 +67,6 @@ impl TryFrom<&dyn TableLike> for BuildpackConfig {
 
 #[derive(Debug)]
 pub(crate) enum ConfigError {
-    CheckExists(PathBuf, std::io::Error),
     ReadConfig(PathBuf, std::io::Error),
     ParseConfig(PathBuf, ParseConfigError),
 }
@@ -89,8 +74,9 @@ pub(crate) enum ConfigError {
 #[derive(Debug)]
 pub(crate) enum ParseConfigError {
     InvalidToml(toml_edit::TomlError),
+    MissingNamespacedConfig,
+    ParseRequestedPackage(Box<ParseRequestedPackageError>),
     WrongConfigType,
-    ParseRequestedPackage(ParseRequestedPackageError),
 }
 
 impl From<ConfigError> for DebianPackagesBuildpackError {
@@ -103,6 +89,32 @@ impl From<ConfigError> for libcnb::Error<DebianPackagesBuildpackError> {
     fn from(value: ConfigError) -> Self {
         Self::BuildpackError(value.into())
     }
+}
+
+fn read_config_file(config_file: impl AsRef<Path>) -> Result<String, ConfigError> {
+    let config_file = config_file.as_ref();
+    fs::read_to_string(config_file)
+        .map_err(|e| ConfigError::ReadConfig(config_file.to_path_buf(), e))
+}
+
+fn parse_config_toml(config: &str) -> Result<DocumentMut, ParseConfigError> {
+    DocumentMut::from_str(config).map_err(ParseConfigError::InvalidToml)
+}
+
+fn get_buildpack_namespaced_config(doc: &DocumentMut) -> Result<&dyn TableLike, ParseConfigError> {
+    let mut current_table = doc
+        .as_item()
+        .as_table_like()
+        .ok_or(ParseConfigError::WrongConfigType)?;
+    for name in NAMESPACED_CONFIG.split('.') {
+        current_table = match current_table.get(name) {
+            Some(item) => item
+                .as_table_like()
+                .ok_or(ParseConfigError::WrongConfigType),
+            None => Err(ParseConfigError::MissingNamespacedConfig),
+        }?;
+    }
+    Ok(current_table)
 }
 
 #[cfg(test)]
@@ -171,8 +183,10 @@ schema-version = "0.2"
 schema-version = "0.2"
         "#
         .trim();
-        let config = BuildpackConfig::from_str(toml).unwrap();
-        assert_eq!(config, BuildpackConfig::default());
+        match BuildpackConfig::from_str(toml).unwrap_err() {
+            ParseConfigError::MissingNamespacedConfig => {}
+            e => panic!("Not the expected error - {e:?}"),
+        }
     }
 
     #[test]