+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Detect now passes if an `Aptfile` exists and will print a migration message during the build phase.
+
+### Changed
+
+- Detect now passes if `project.toml` is present and contains namespaced configuration.
+
 ## [0.0.3] - 2024-12-05
 
 ### Changed
@@ -37,6 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release.
 
 [unreleased]: https://github.com/heroku/buildpacks-deb-packages/compare/v0.0.3...HEAD
+
 [0.0.3]: https://github.com/heroku/buildpacks-deb-packages/compare/v0.0.2...v0.0.3
+
 [0.0.2]: https://github.com/heroku/buildpacks-deb-packages/compare/v0.0.1...v0.0.2
+
 [0.0.1]: https://github.com/heroku/buildpacks-deb-packages/releases/tag/v0.0.1